Cross-platform Makefile, source reorg into files/ and user/, fix all warnings
Makefile: rewrite with platform detection (Darwin/Linux), build configs
(release default, DEBUG=1), quiet builds (V=1 for verbose), install target,
help target, AR variable, LDFLAGS/LDLIBS separation, tool warning suppression.

Source reorg: move file transfer code to src/files/ (libfiles.a), user
management to src/user/ (libuser.a). Update include paths across 30+ files.

Portable forkpty: console.cpp uses <util.h> on macOS, <pty.h> on Linux.

Fix 54 compiler warnings across 26 files:
- 7 real bugs: dangling pointers used as buffers (lilo.cpp, personal.cpp,
  config.cpp), return-stack-address (msgbase.cpp, wfc.cpp), operator
  precedence (lilo.cpp !x==255), wrong loop variable (diredit.cpp),
  off-by-one array writes (lilo.cpp, config.cpp, file3.cpp), unsequenced
  modification (stringed.cpp), uninitialized double (file.cpp)
- 25 sign-compare warnings (int vs size_t casts)
- 5 empty-body if(x); → (void)x
- 4 constant-conversion 196→(char)196
- Unused functions/variables, self-assign, tautological compares

+++ src/cmd_registry.cpp
@@ -36,16 +36,16 @@
 #include "utility.h"
 #include "jam_bbs.h"
 #include "msgbase.h"
-#include "file.h"
-#include "file1.h"
-#include "file2.h"
-#include "file3.h"
-#include "archive.h"
-#include "filesys.h"
+#include "files/file.h"
+#include "files/file1.h"
+#include "files/file2.h"
+#include "files/file3.h"
+#include "files/archive.h"
+#include "files/filesys.h"
 #include "config.h"
-#include "diredit.h"
-#include "newuser.h"
-#include "uedit.h"
+#include "files/diredit.h"
+#include "user/newuser.h"
+#include "user/uedit.h"
 #include "mm1.h"
 #include "stringed.h"
 #include "session.h"
@@ -451,7 +451,7 @@ void othercmd(char type,char ms[40])
         selfValidationCheck(ms);
         break;
     case ';':
-        for(c=0;c<strlen(ms);c++) {
+        for(c=0;c<(int)strlen(ms);c++) {
             if(ms[c]==';') ms[c]=13;
         }
         ms[strlen(ms)+1]=13;
